authorization refactor: move to src folder
