+++ src/interfaces/authInterfaces/webservice.go
@@ -1,7 +1,7 @@
 package authInterfaces
 
 import (
-	"yurko/usecases/authUsecases"
+	"usecases/authUsecases"
 )
 
 type RegistrationInteractor interface {
