+++ src/interfaces/index.go
@@ -1,6 +1,6 @@
 package interfaces
 
-import "yurko/infrastructures"
+import "infrastructures"
 
 func GetCurrentUser(sessionId string) string {
 	return infrastructures.InMemorySession.Get(sessionId)
